requirements.txt
mysql-connector-python
SQLAlchemy
pymysql
pandas
streamlit
faker
